+++ jest.setup.ts
@@ -1,5 +1,5 @@
 import { setupServer } from 'msw/node';
-import { handlers } from './__mocks__/provider/handlers';
+import { handlers } from './__mocks__/handlers';
 
 const server = setupServer(...handlers);
